+++ EnvelopeBlock.sln
@@ -1,18 +1,37 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-# SharpDevelop 4.1.0.8000
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EnvelopeBlock", "EnvelopeBlock\EnvelopeBlock.csproj", "{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}"
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35327.3
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnvelopeBlock", "EnvelopeBlock\EnvelopeBlock.csproj", "{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|x86.Build.0 = Debug|x86
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|x64.ActiveCfg = Debug|x64
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|x64.Build.0 = Debug|x64
 		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|x86.ActiveCfg = Debug|x86
-		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|x86.Build.0 = Release|x86
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Debug|x86.Build.0 = Debug|x86
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|x64.ActiveCfg = Release|x64
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|x64.Build.0 = Release|x64
 		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|x86.ActiveCfg = Release|x86
+		{2837EBA5-2F9A-4192-A2E7-95D41F6C478F}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7CBC0C27-2A83-4865-AC08-3E73A12BA041}
 	EndGlobalSection
 EndGlobal
