+++ EnvelopeBlock/EnvelopeBlockEvent.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel;
+
+namespace EnvelopeBlock
+{
+    public class EnvelopeBlockEvent : INotifyPropertyChanged
+    {
+        public int envelopPatternIndex;
+        public event PropertyChangedEventHandler PropertyChanged;
+    }
+}
