+++ EnvelopeBlock/IEnvelopePoint.cs
@@ -1,0 +1,11 @@
+﻿namespace EnvelopeBlock
+{
+    /// <summary>
+    /// Helper interface
+    /// </summary>
+    public interface IEnvelopePoint
+    {
+        double TimeStamp { get; set; }
+        double Value { get; set; }
+    }
+}
